+++ packages/toolkit/src/vin/index.ts
@@ -1,3 +1,5 @@
 export * from './validate';
 export * from './types';
 export * from './constants';
+export { InvalidVINError } from './errors';
+export { parseVIN } from './parse';